Mes 3659 restructure folders for cat b debrief page (#828)
* mes-3659-restructuredDebriefForCatB

* mes-3659-addedPageToPageConstants
