+++ src/pages/debrief/cat-b/debrief.cat-b.page.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from '@angular/core';
+import { IonicPageModule } from 'ionic-angular';
+import { DebriefCatBPage } from './debrief.cat-b.page';
+import { EffectsModule } from '@ngrx/effects';
+import { DebriefAnalyticsEffects } from '../debrief.analytics.effects';
+import { ComponentsModule } from '../../../components/common/common-components.module';
+import { DebriefComponentsModule } from '../components/debrief-components.module';
+import { TranslateModule } from 'ng2-translate';
+import { DebriefEffects } from '../debrief.effects';
+
+@NgModule({
+  declarations: [
+    DebriefCatBPage,
+  ],
+  imports: [
+    DebriefComponentsModule,
+    IonicPageModule.forChild(DebriefCatBPage),
+    EffectsModule.forFeature([
+      DebriefEffects,
+      DebriefAnalyticsEffects,
+    ]),
+    ComponentsModule,
+    TranslateModule,
+  ],
+})
+export class DebriefCatBPageModule { }
