+++ src/pages/debrief/cat-b/debrief.cat-b.page.scss
@@ -1,4 +1,4 @@
-page-debrief > * {
+debrief-cat-b-page > * {
   #test-outcome-background {
     text-align: center;
     padding: 16px 0;
